Stop numerous sqlite databases from being created
This makes a change so that a new sqlite database isn't created for
every test case. There are currently issues with disconnecting from
sessions in the beta version of the SQL adapter and these are
preventing the test database file from being able to be deleted
after the tests have completed.