Move relational fields out of abstract base models
ChatterBot's abstract base models for Django are suppose to make
it easier for a developer to create a new model, based on the
existing ones with their own modifications.

Unfortunately, having related fields like foregin keys will result
in errors when trying to create new models because the fields of
the abstract parent class are still evaluated by Django and
because they can only reference the origional models.